serverless/requirements.txt
backports.shutil-get-terminal-size==1.0.0
botocore==1.4.65
chalice==0.3.0
click==6.2
decorator==4.0.10
dnspython==1.15.0
docutils==0.12
enum34==1.1.6
httpie==0.9.6
ipaddr==2.1.11
ipwhois==0.14.0
ipython==5.1.0
ipython-genutils==0.1.0
jmespath==0.9.0
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.8
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
python-geoip==1.2
python-geoip-geolite2==2015.303
requests==2.11.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
typing==3.5.1.0
virtualenv==15.0.3
wcwidth==0.1.7
whois==0.7
